+++ gpflow/conditionals/util.py
@@ -542,3 +542,54 @@ def mix_latent_gp(W, g_mean, g_var, full_cov, full_output_cov):
     tf.debugging.assert_shapes(shape_constraints, message="mix_latent_gp()")
 
     return f_mean, f_var
+
+
+def separate_independent_conditional_implementation(
+    Kmns, Kmms, Knns, f, *, full_cov=False, full_output_cov=False, q_sqrt=None, white=False,
+):
+    """Multi-output GP with independent GP priors.
+    Number of latent processes equals the number of outputs (L = P).
+    The covariance matrices used to calculate the conditional have the following shape:
+    - Kuu: [P, M, M]
+    - Kuf: [P, M, N]
+    - Kff: [P, N] or [P, N, N]
+
+    Further reference
+    -----------------
+    - See `gpflow.conditionals._conditional` for a detailed explanation of
+      conditional in the single-output case.
+    - See the multioutput notebook for more information about the multioutput framework.
+    - See above for the parameters and the return value.
+    """
+    fs = tf.transpose(f)[:, :, None]  # [P, M, 1]
+    # [P, 1, M, M]  or  [P, M, 1]
+
+    if q_sqrt is not None:
+        q_sqrts = (
+            tf.transpose(q_sqrt)[:, :, None] if q_sqrt.shape.ndims == 2 else q_sqrt[:, None, :, :]
+        )
+        base_conditional_args_to_map = (Kmms, Kmns, Knns, fs, q_sqrts)
+
+        def single_gp_conditional(t):
+            Kmm, Kmn, Knn, f, q_sqrt = t
+            return base_conditional(Kmn, Kmm, Knn, f, full_cov=full_cov, q_sqrt=q_sqrt, white=white)
+
+    else:
+        base_conditional_args_to_map = (Kmms, Kmns, Knns, fs)
+
+        def single_gp_conditional(t):
+            Kmm, Kmn, Knn, f = t
+            return base_conditional(Kmn, Kmm, Knn, f, full_cov=full_cov, q_sqrt=q_sqrt, white=white)
+
+    rmu, rvar = tf.map_fn(
+        single_gp_conditional, base_conditional_args_to_map, (default_float(), default_float())
+    )  # [P, N, 1], [P, 1, N, N] or [P, N, 1]
+
+    fmu = rollaxis_left(tf.squeeze(rmu, axis=-1), 1)  # [N, P]
+
+    if full_cov:
+        fvar = tf.squeeze(rvar, axis=-3)  # [..., 0, :, :]  # [P, N, N]
+    else:
+        fvar = rollaxis_left(tf.squeeze(rvar, axis=-1), 1)  # [N, P]
+
+    return fmu, expand_independent_outputs(fvar, full_cov, full_output_cov)
